requirements.txt
numpy==1.22.1
pandas==1.3.5
requests==2.27.1
psycopg2-binary
SQLAlchemy==1.4.30
